Fixed the popup so it behaves uniformly. Added back-button UX.
Added one type of Fish (yellow), added a primaryColor attribute to the fish. This makes it so that fish can get a text-color of their primary color for example. Added functionality to the back button so it exits the app if the user clicks on it twice.  Removed the ability to dismiss the popup-window by clicking on the popup in CaughtActivity.