Fixed code-debt, refactored reel.
Reel is now separated into the two different activities, as to make more sense when looking at the class-names. What was in reel that was related to the throw-activity is now in the ThrowActivity class. The ThrowActivity class had some small fixes. The Reel-class is now separated into more logical functions. Added initializeScreen as a helper-method. Added a method to always return the rounded distance instead of the 7 decimals.