Disabled CC rotating the reel. Small fixes.
Had to change GameActivities to an AppCompatActivity. The reel-image was not shown otherwise. No idea why but was a quick fix. Also disabled counter-clockwise rotation of the reel as that would increase the distance. It does not behave exactly as I want it to just yet. Will require to use some kind of linear algebra to solve.