Refactoring, made language uniform
Made language uniform in English over all menus. Also changed all instruction-texts to have a shadow as well as be the same font-size and roughly in the same position. These changes work in both light-mode and dark-mode and look uniform in both modes.
Also made sure that the phone cannot sleep while the game is running, as well as a navigation back from the tutorial. Refactored and removed some unused files after all of this was done.